third_party/WebKit/LayoutTests/platform/win/http/tests/security/contentSecurityPolicy/require-sri-for/require-sri-for-style-reportonly-blocked-expected.txt

CONSOLE ERROR: The Content Security Policy 'require-sri-for style;' was delivered in report-only mode, but does not specify a 'report-uri'; the policy will have no effect. Please either add a 'report-uri' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE ERROR: line 15: [Report Only] Refused to load the stylesheet 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/style-set-red.css' because 'require-sri-for' directive requires integrity attribute be present for all stylesheets.
This is a testharness.js-based test.
PASS Stylesheets without integrity generate reports. 
PASS Stylesheets without integrity do not load. 
Harness: the test ran to completion.
